backend/requirements.txt
Flask
flask-cors
transformers
torch
python-dotenv
accelerate